requirements.txt
flask
flask-socketio
eventlet
flask_sqlalchemy
psycopg2-binary
python-dotenv